HBX-1625: Move class 'org.hibernate.cfg.reveng.DefaultDatabaseCollector' to 'org.hibernate.tool.internal.reveng.DefaultDatabaseCollector'
