+++ test/oracle/src/main/java/org/hibernate/cfg/reveng/dialect/TestCase.java
@@ -3,13 +3,13 @@ package org.hibernate.cfg.reveng.dialect;
 import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
 import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.cfg.Environment;
-import org.hibernate.cfg.reveng.DefaultDatabaseCollector;
 import org.hibernate.engine.jdbc.connections.spi.ConnectionProvider;
 import org.hibernate.engine.jdbc.spi.JdbcServices;
 import org.hibernate.service.ServiceRegistry;
 import org.hibernate.tool.api.dialect.MetaDataDialect;
 import org.hibernate.tool.api.reveng.ReverseEngineeringRuntimeInfo;
 import org.hibernate.tool.internal.dialect.OracleMetaDataDialect;
+import org.hibernate.tool.internal.metadata.DefaultDatabaseCollector;
 import org.hibernate.tools.test.util.JdbcUtil;
 import org.junit.After;
 import org.junit.Before;
